Moved Cell and CellStates into the Entity System.
