+++ conway_module/lib/entity-system/CellStates.js
@@ -1,7 +1,9 @@
 /**
  * Defines the possible states a Cell can have.
  */
-module.exports.CellStates = {
+const CellStates = {
 	DEAD: 0,
 	ALIVE: 1,
 };
+
+module.exports = CellStates;
