+++ conway_module/test/core/QuadTreeTestHelper.js
@@ -1,4 +1,4 @@
-const { Cell } = require('./../../lib/core/Quadtree.js');
+const { Cell } = require('./../../lib/entity-system/Entities.js');
 /*
 			 0 1 2 3 4 5 6 7 8 9
 		0 |1|0|0|0|0|0|0|0|0|0|
